+++ app/Models/Lead.php
@@ -9,6 +9,7 @@ use App\Traits\HasCompany;
 use Illuminate\Database\Eloquent\Casts\Attribute;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
+use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Notifications\Notifiable;
 
 /**
@@ -119,7 +120,7 @@ class Lead extends BaseModel
 
     public function getImageUrlAttribute()
     {
-        $gravatarHash = !is_null($this->email) ? md5(strtolower(trim($this->email))) : '';
+        $gravatarHash = !is_null($this->client_email) ? md5(strtolower(trim($this->client_email))) : '';
 
         return 'https://www.gravatar.com/avatar/' . $gravatarHash . '.png?s=200&d=mp';
     }
@@ -140,7 +141,7 @@ class Lead extends BaseModel
     // phpcs:ignore
     public function routeNotificationForMail($notification)
     {
-        return $this->email;
+        return $this->client_email;
     }
 
     public function leadSource(): BelongsTo
@@ -208,6 +209,11 @@ class Lead extends BaseModel
 
         // Retrieve leads
         return $leadsQuery->get();
+    }
+
+    public function communicationActivities(): HasMany
+    {
+        return $this->hasMany(CommunicationActivity::class, 'lead_id')->orderByDesc('timestamp');
     }
 
 }