+++ resources/views/leads/ajax/show.blade.php
@@ -12,7 +12,7 @@ $viewLeadFollowupPermission = user()->permission('view_lead_follow_up');
 
     <div class="row">
         <!--  USER CARDS START -->
-        <div class="col-sm-9 mb-4 mb-xl-0 mb-lg-4 mb-md-0">
+        <div class="col-sm-7 mb-4 mb-xl-0 mb-lg-4 mb-md-0">
 
             <x-cards.data :title="__('modules.deal.dealInfo')">
 
@@ -171,33 +171,8 @@ $viewLeadFollowupPermission = user()->permission('view_lead_follow_up');
         </div>
         <!--  USER CARDS END -->
 
-        <div class="col-sm-3">
-
-
-            {{-- <x-cards.data :title="__('modules.leadContact.leadDetails')">
-
-                <x-cards.data-row :label="__('modules.leadContact.leadContact')" otherClasses="pr-1" labelClasses="pr-1"
-                                    value="<a href='{{ route('lead-contact.show', $deal->contact->id) }}' class='text-darkest-grey'> {{ $deal->contact->client_name_salutation }}</a>"/>
-
-                <x-cards.data-row :label="__('app.email')" :value="$deal->contact->client_email ?? '--'" otherClasses="pr-1" labelClasses="pr-1"/>
-                <x-cards.data-row :label="__('modules.lead.mobile')" :value="$deal->contact->mobile ?? '--'" otherClasses="pr-1" labelClasses="pr-1"/>
-
-                <x-cards.data-row :label="__('modules.lead.companyName')"
-                                    :value="!empty($deal->contact->company_name) ? $deal->contact->company_name : '--'" otherClasses="pr-1" labelClasses="pr-1"/>
-
-                <div class="d-flex">
-                    @if ($deal->contact->client_email)
-                        <x-forms.link-secondary class="mr-3 pr-1" link='mailto:{{ $deal->contact->client_email }}'
-                                                icon="envelope">@lang('app.email')</x-forms.link-secondary>
-                    @endif
-
-                    @if ($deal->contact->mobile )
-                        <x-forms.button-secondary class="btn-copy pr-1" data-clipboard-text="{{ $deal->contact->mobile }}"
-                                                    icon="phone">@lang('app.mobile')</x-forms.button-secondary>
-                    @endif
-                </div>
-
-            </x-cards.data> --}}
+        <div class="col-sm-5">
+            @include('leads.ajax.activities')
         </div>
     </div>